requirements.txt
requests
python-dotenv
watchdog
python-telegram-bot